+++ lib/osm/event.rb
@@ -37,6 +37,8 @@ module Osm
     #   @return [Fixnum] the maximum number of people who can attend the event (0 = no limit)
     # @!attendance [rw] attendance_limit_includes_leaders
     #   @return [Boolean] whether the attendance limit includes leaders
+    # @!attribute [rw] attendance_reminder
+    #   @return [Fixnum] how many days before the event to send a reminder to those attending (0 (off), 1, 3, 7, 14, 21, 28)
     # @!attribute [rw] allow_booking
     #   @return [Boolean] whether booking is allowed through My.SCOUT
 
@@ -57,11 +59,13 @@ module Osm
     attribute :reminders, :type => Boolean, :default => true
     attribute :attendance_limit, :type => Integer, :default => 0
     attribute :attendance_limit_includes_leaders, :type => Boolean, :default => false
+    attribute :attendance_reminder, :type => Integer, :default => 0
     attribute :allow_booking, :type => Boolean, :default => true
 
     attr_accessible :id, :section_id, :name, :start, :finish, :cost, :location, :notes, :archived,
                     :fields, :columns, :notepad, :public_notepad, :confirm_by_date, :allow_changes,
-                    :reminders, :attendance_limit, :attendance_limit_includes_leaders, :allow_booking
+                    :reminders, :attendance_limit, :attendance_limit_includes_leaders,
+                    :attendance_reminder, :allow_booking
 
     validates_numericality_of :id, :only_integer=>true, :greater_than=>0, :allow_nil => true
     validates_numericality_of :section_id, :only_integer=>true, :greater_than=>0
@@ -72,6 +76,7 @@ module Osm
     validates_inclusion_of :reminders, :in => [true, false]
     validates_inclusion_of :attendance_limit_includes_leaders, :in => [true, false]
     validates_inclusion_of :allow_booking, :in => [true, false]
+    validates_inclusion_of :attendance_reminder, :in => [0, 1, 3, 7, 14, 21, 28]
     validates_format_of :cost, :with => /\A(?:\d+\.\d{2}|TBC)\Z/
 
 
@@ -162,6 +167,7 @@ module Osm
         'allowChanges' => event.allow_changes ? 'true' : 'false',
         'disablereminders' => !event.reminders ? 'true' : 'false',
         'attendancelimit' => event.attendance_limit,
+        'attendancereminder' => event.attendance_reminder,
         'limitincludesleaders' => event.attendance_limit_includes_leaders ? 'true' : 'false',
         'allowbooking' => event.allow_booking ? 'true' : 'false',
       })
@@ -200,6 +206,7 @@ module Osm
         'allowChanges' => allow_changes ? 'true' : 'false',
         'disablereminders' => !reminders ? 'true' : 'false',
         'attendancelimit' => attendance_limit,
+        'attendancereminder' => attendance_reminder,
         'limitincludesleaders' => attendance_limit_includes_leaders ? 'true' : 'false',
         'allowbooking' => allow_booking ? 'true' : 'false',
       })
@@ -392,6 +399,7 @@ module Osm
         :reminders => !event_data['disablereminders'].eql?('1'),
         :attendance_limit => event_data['attendancelimit'].to_i,
         :attendance_limit_includes_leaders => event_data['limitincludesleaders'].eql?('1'),
+        :attendance_reminder => event_data['attendancereminder'].to_i,
         :allow_booking => event_data['allowbooking'].eql?('1'),
       )